# Pool histograms for reuse between active and inactive