<!-- Thanks for sending a pull request!

If this is your first contribution, please review and sign our contributor agreement -
https://www.elastic.co/contributor-agreement.

Guidelines:
 - Prefer small PRs, and split changes into multiple logical commits where they must
   be delivered in a single PR.
 - If the PR is incomplete and not yet ready for review, open it as a Draft.
 - Once the PR is marked ready for review it is expected to pass all tests and linting,
   and you should not force-push any changes.

See also https://github.com/elastic/apm-server/blob/main/CONTRIBUTING.md for more tips on contributing.
-->

## Motivation/summary

<!--
Describe your change in the title and description, and provide a motivation for the
change and rationale for the approach taken.
-->
Pool histograms for reuse between active and inactive. This should cut txmetrics and servicetxmetrics histograms memory usage in half.

### Benchmark

**TLDR: max_rss 7857737728 -> 4583124992**

Before:
```
Benchmark10000AggregationGroups-512          609        1151431598 ns/op                 0 error_responses/sec           0 errors/sec        13034 events/sec           87.00 gc_cycles        333.0 max_goroutines     7343182624 max_heap_alloc         37562153 max_heap_objects     7153254400 max_rss              16.09 mean_available_indexers          1525 metrics/sec        6391 spans/sec          5118 txs/sec      282429347 B/op   3339987 allocs/op
Benchmark10000AggregationGroups-512           84        1551815749 ns/op                 0 error_responses/sec           0 errors/sec        10411 events/sec            8.000 gc_cycles               121.0 max_goroutines     8013441408 max_heap_alloc         39208667 max_heap_objects     7670460416 max_rss              17.62 mean_available_indexers          1364 metrics/sec        4974 spans/sec          4074 txs/sec      296773386 B/op   3534176 allocs/op
Benchmark10000AggregationGroups-512           97        1525567700 ns/op                 0 error_responses/sec           0 errors/sec        10522 events/sec            9.000 gc_cycles               134.0 max_goroutines     7926363520 max_heap_alloc         41798615 max_heap_objects     7857737728 max_rss              17.56 mean_available_indexers          1167 metrics/sec        5094 spans/sec          4261 txs/sec      295408565 B/op   3545122 allocs/op
```

After:
```
Benchmark10000AggregationGroups-512          102        1585284156 ns/op                 0 error_responses/sec           0 errors/sec         9823 events/sec           16.00 gc_cycles        130.0 max_goroutines     4595239120 max_heap_alloc         24614495 max_heap_objects     4044644352 max_rss              17.76 mean_available_indexers          1059 metrics/sec        4846 spans/sec          3918 txs/sec      280151177 B/op   3473866 allocs/op
Benchmark10000AggregationGroups-512          102        1600855506 ns/op                 0 error_responses/sec           0 errors/sec        10229 events/sec           15.00 gc_cycles        133.0 max_goroutines     5217648352 max_heap_alloc         26529472 max_heap_objects     4897488896 max_rss              17.87 mean_available_indexers          1540 metrics/sec        4803 spans/sec          3885 txs/sec      286485461 B/op   3537505 allocs/op
Benchmark10000AggregationGroups-512           79        1519381483 ns/op                 0 error_responses/sec           0 errors/sec        10575 events/sec           12.00 gc_cycles        121.0 max_goroutines     4667991616 max_heap_alloc         25253503 max_heap_objects     4583124992 max_rss              17.76 mean_available_indexers          1263 metrics/sec        5087 spans/sec          4225 txs/sec      286105520 B/op   3531890 allocs/op

```

## Checklist

<!--
Delete irrelevant items. The changelog should only be updated for user-facing changes.
Once the PR is ready for review there should be no unticked boxes.
-->

- [x] Update [CHANGELOG.asciidoc](https://github.com/elastic/apm-server/blob/main/CHANGELOG.asciidoc)
~~- [ ] Update [package changelog.yml](https://github.com/elastic/apm-server/blob/main/apmpackage/apm/changelog.yml) (only if changes to `apmpackage` have been made)~~
~~- [ ] Documentation has been updated~~

## How to test these changes

<!--
Explain how this PR can be tested by the reviewer: commands, dependencies, steps, etc.
If it is self-explanatory, delete this section.
-->
N/A

## Related issues

<!--
Reference the related issue(s), and make use of magic keywords where it makes sense
https://help.github.com/articles/closing-issues-using-keywords/.
-->
Part of #10147